+++ src/OpcUaStackServer/InformationModel/MethodMap.h
@@ -21,7 +21,7 @@
 #include <map>
 
 #include "OpcUaStackServer/ServiceSetApplication/ForwardMethodSync.h"
-#include "OpcUaStackCore/ServiceSetApplication/ForwardMethodAsync.h"
+#include "OpcUaStackServer/ServiceSetApplication/ForwardMethodAsync.h"
 #include "OpcUaStackServer/InformationModel/MethodId.h"
 
 namespace OpcUaStackServer
@@ -31,7 +31,7 @@ namespace OpcUaStackServer
 	{
 	  public:
 		typedef std::map<MethodId, ForwardMethodSync::SPtr> ForwardMethodSyncMap;
-		typedef std::map<MethodId, OpcUaStackCore::ForwardMethodAsync::SPtr> ForwardMethodAsyncMap;
+		typedef std::map<MethodId, ForwardMethodAsync::SPtr> ForwardMethodAsyncMap;
 
 		MethodMap(
 			void
@@ -59,7 +59,7 @@ namespace OpcUaStackServer
 		bool registerMethodAsync(
 			const OpcUaStackCore::OpcUaNodeId& objectNodeId,
 			const OpcUaStackCore::OpcUaNodeId& methodNodeId,
-			OpcUaStackCore::ForwardMethodAsync::SPtr& forwardMethodAsync
+			ForwardMethodAsync::SPtr& forwardMethodAsync
 		);
 		bool deregisterMethodSync(
 			const OpcUaStackCore::OpcUaNodeId& objectNodeId,
@@ -73,7 +73,7 @@ namespace OpcUaStackServer
 			const OpcUaStackCore::OpcUaNodeId& objectNodeId,
 			const OpcUaStackCore::OpcUaNodeId& methodNodeId
 		);
-		OpcUaStackCore::ForwardMethodAsync::SPtr getMethodAsync(
+		ForwardMethodAsync::SPtr getMethodAsync(
 			const OpcUaStackCore::OpcUaNodeId& objectNodeId,
 			const OpcUaStackCore::OpcUaNodeId& methodNodeId
 		);
